+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Quest | Added workflow configurations
 - EIR | Added workflow configurations
 - ANC | Report Measures List Items, Filter Selection
+- ANC | Report Measures List Items, Filter Selection
 - Quest | Add photo capture questionnaire widget
 - Quest | Add ability to edit questionnaire responses
 
@@ -53,7 +54,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - ANC | Bottom sheet for switching registers added
 - ANC | User actions, profile screen implemented
 - ANC | Drawer menu removed
-- ANC | Reports List UI Items
 
 ### Fixed